feat: Permit optional pass through of JSON as string
If the JSON retrieved from the database is simply to be re-serialized as a string, parsing
it into a javascript object is a waste of resources.
The default JSON parser does not support high numeric precision for JSON numbers, whereas
the JSONP format in postgres does. Using a value transformer with the new passthrough option
permits the use of a custom parser (e.g. json-bigint)

Closes #8128